requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
websockets==15.0.1
python-multipart==0.0.20
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
python-dotenv==1.1.1 